feat: social for linked to social media

+++ admin/social/index.php
@@ -1,12 +1,12 @@
 <?php
 include_once('../koneksi.php');
 $no = 1;
+
 $query = mysqli_query(
   $koneksi,
-  "SELECT tb_social.*, tb_icons.icons
-  FROM tb_social
-  INNER JOIN tb_icons
-  ON tb_social.id_icons = tb_icons.id
+  "SELECT tb_social.*, tb_icons.icon 
+  FROM tb_social 
+  INNER JOIN tb_icons ON tb_social.icon = tb_icons.id 
   ORDER BY id DESC"
 );
 
@@ -33,42 +33,29 @@ $query = mysqli_query(
           </tr>
         </thead>
         <tbody>
-          <?php
-
-          while ($data = mysqli_fetch_array($query)) {
-          ?>
+          <?php while ($data = mysqli_fetch_array($query)) : ?>
             <tr>
               <td><?= $no++; ?></td>
-              <td><?= $data['nama_sosmed']; ?></td>
+              <td><?= $data['name_social']; ?></td>
               <td class="text-center">
-                <i style="border: 1px solid #000; border-radius: 5px;" class="text-xl p-3 fas far fa fa-<?= $data['icons']; ?>"></i>
+                <i style="border: 1px solid #000; border-radius: 5px;" class="text-xl p-3 fas far fa fa-<?= $data['icon']; ?>"></i>
               </td>
 
               <td><?= $data['url']; ?></td>
               <td class="text-center">
                 <a href="social/edit.php?id=<?= $data['id']; ?>&page=social" class="btn btn-warning text-light">Edit</a>
-                <a href="social/proses_hapus.php?id=<?= $data['id']; ?>&page=social" class="btn btn-danger">Hapus</a>
+                <a onclick="return confirm('Apakah anda yakin ingin menghapusnya?')" href="social/proses_hapus.php?id=<?= $data['id']; ?>&page=social" class="btn btn-danger">Hapus</a>
               </td>
             </tr>
-          <?php }
 
-          ?>
+          <?php endwhile; ?>
         </tbody>
       </table>
     </div>
     <!-- /.card-body -->
     <div class="card-footer clearfix">
+      <p>For reference icon see from this <a href="https://adminlte.io/themes/AdminLTE/pages/UI/icons.html">Link</a></p>
     </div>
   </div>
 </section>
 <!-- MAIN CONTENT -->
-
-
-<!-- Key -->
-
-<!--                   
-  <div class="col-md-3 col-sm-4">
-    <i class="fas fa fa-facebook"></i>
-    <i class="fas fa fa-chrome"></i>
-  </div> 
--->
